Education sectoral composite - add score and dummy for in need

Education sectoral composite - add score and dummy for in need

Usage

add_comp_edu(
  df,
  ind_schooling_age_d_n = "edu_ind_schooling_age_d_n",
  ind_no_access_d_n = "edu_ind_no_access_d_n",
  ind_barrier_protection_d_n = "ind_barrier_protection_d_n",
  occupation_d_n = "edu_disrupted_occupation_d_n",
  hazards_d_n = "edu_disrupted_hazards_d_n",
  displaced_d_n = "edu_disrupted_displaced_d_n",
  teacher_d_n = "edu_disrupted_teacher_d_n"
)

Arguments

df

A data frame.

ind_schooling_age_d_n

Column name for the number of children of schooling age.

ind_no_access_d_n

Column name for the number of children with no access to education.

ind_barrier_protection_d_n

Column name for the number of children with barriers to protection.

occupation_d_n

Column name for the number of children with disrupted education due to the school being occupied by armed groups

hazards_d_n

Column name for the number of children with disrupted education due to hazards.

displaced_d_n

Column name for the number of children with disrupted education due to a recent displacement.

teacher_d_n

Column name for the number of children with disrupted education due to teachers' absence.