Add income source amount as proportions of total income

Add income source amount as proportions of total income

Usage

add_income_source_prop(
  df,
  income_souce_salaried_n = "cm_income_source_salaried_n",
  income_source_casual_n = "cm_income_source_casual_n",
  income_source_own_business_n = "cm_income_source_own_business_n",
  income_source_own_production_n = "cm_income_source_own_production_n",
  income_source_social_benefits_n = "cm_income_source_social_benefits_n",
  income_source_rent_n = "cm_income_source_rent_n",
  income_source_remittances_n = "cm_income_source_remittances_n",
  income_source_assistance_n = "cm_income_source_assistance_n",
  income_source_support_friends_n = "cm_income_source_support_friends_n",
  income_source_donation_n = "cm_income_source_donation_n",
  income_source_other_n = "cm_income_source_other_n"
)

Arguments

df

A data frame

income_souce_salaried_n

Column name for salaried income amount.

income_source_casual_n

Column name for casual income amount.

income_source_own_business_n

Column name for own business income amount.

income_source_own_production_n

Column name for own production income amount.

income_source_social_benefits_n

Column name for social benefits income amount.

income_source_rent_n

Column name for rent income amount.

income_source_remittances_n

Column name for remittances income amount.

income_source_assistance_n

Column name for assistance income amount.

income_source_support_friends_n

Column name for support from friends income amount.

income_source_donation_n

Column name for donation income amount.

income_source_other_n

Column name for other income amount.