FIES - Food Insecurity Experience Scale analysis

fies_raw_score() calculates the raw score of the FIES. For parameter estimates and assessment statistics with a Rasch model thanks to package RM.weights, see impactR.analysis::fies_estimates() (under construction).

Usage

fies_raw_score(
  df,
  fies_worried = "fies_worried",
  fies_healthy = "fies_healthy",
  fies_fewfoods = "fies_fewfoods",
  fies_skipped = "fies_skipped",
  fies_ateless = "fies_ateless",
  fies_ranout = "fies_ranout",
  fies_hungry = "fies_hungry",
  fies_wholeday = "fies_wholeday",
  level_codes = c("yes", "no", "do_not_know", "prefer_not_to_answer")
)

Arguments

df

A data frame

fies_worried

Component column: Worried about not having enough food to eat.

fies_healthy

Component column: Unable to eat healthy or nutritious food.

fies_fewfoods

Component column: Ate only a few kinds of foods.

fies_skipped

Component column: Skip a meal.

fies_ateless

Component column: Ate less than thought to should.

fies_ranout

Component column: Ran out of food.

fies_hungry

Component column: Hungry without eating.

fies_wholeday

Component column: No eating for a whole day.

level_codes

Character vector of responses codes, including first in the following order: "Yes", "No", "Do not know", "Prefer not to answer", e.g., c("yes", "no", "do_not_know", "prefer_not_to_answer")

Value

Nine new columns: each component dummy variable if "Yes" (fies_*_d) and FIES raw score (fiew_raw_score).

Details

THe recall period can be different (30 days or 12 months) depending on the research's objectives (food insecurity or population monitoring for instance).