Education sectoral composite - add score and dummy for in need

Education sectoral composite - add score and dummy for in need

Usage

add_comp_edu(
  df,
  schooling_age_n = "edu_schooling_age_n",
  no_access_n = "edu_no_access_n",
  barrier_protection_n = "edu_barrier_protection_n",
  occupation_n = "edu_disrupted_occupation_n",
  hazards_n = "edu_disrupted_hazards_n",
  displaced_n = "edu_disrupted_displaced_n",
  teacher_n = "edu_disrupted_teacher_n"
)

Arguments

df

A data frame.

schooling_age_n

Column name for the number of children of schooling age.

no_access_n

Column name for the number of children with no access to education.

barrier_protection_n

Column name for the number of children with barriers to protection.

occupation_n

Column name for the number of children with disrupted education due to the school being occupied by armed groups

hazards_n

Column name for the number of children with disrupted education due to hazards.

displaced_n

Column name for the number of children with disrupted education due to a recent displacement.

teacher_n

Column name for the number of children with disrupted education due to teachers' absence.