Health composite - add score and dummy for in need

Health composite - add score and dummy for in need

Usage

add_comp_health(
  df,
  ind_healthcare_needed_no_n = "health_ind_healthcare_needed_no_n",
  ind_healthcare_needed_yes_unmet_n = "health_ind_healthcare_needed_yes_unmet_n",
  ind_healthcare_needed_yes_met_n = "health_ind_healthcare_needed_yes_met_n",
  wgq_dis = FALSE,
  ind_healthcare_needed_no_wgq_dis_n = "health_ind_healthcare_needed_no_wgq_dis_n",
  ind_healthcare_needed_yes_unmet_wgq_dis_n =
    "health_ind_healthcare_needed_yes_unmet_wgq_dis_n",
  ind_healthcare_needed_yes_met_wgq_dis_n =
    "health_ind_healthcare_needed_yes_met_wgq_dis_n"
)

Arguments

df

A data frame.

ind_healthcare_needed_no_n

Column name for the number of individuals who did not need to access healthcare.

ind_healthcare_needed_yes_unmet_n

Column name for the number of individuals who needed to access healthcare but did not.

ind_healthcare_needed_yes_met_n

Column name for the number of individuals who needed to access healthcare and did.

wgq_dis

Logical. If TRUE, the function will consider the disability columns.

ind_healthcare_needed_no_wgq_dis_n

Column name for the number of individuals who did not need to access healthcare and have a disability.

ind_healthcare_needed_yes_unmet_wgq_dis_n

Column name for the number of individuals who needed to access healthcare but did not and have a disability.

ind_healthcare_needed_yes_met_wgq_dis_n

Column name for the number of individuals who needed to access healthcare and did and have a disability.