WASH sectoral composite - add score and dummy for in need

WASH sectoral composite - add score and dummy for in need

Usage

add_comp_wash(
  df,
  setting = "setting",
  setting_levels = c("camp", "urban", "rural"),
  drinking_water_quantity = "wash_drinking_water_quantity",
  drinking_water_quantity_levels = c("always", "often", "sometimes", "rarely", "never",
    "dnk", "pnta"),
  drinking_water_quality_jmp_cat = "wash_drinking_water_quality_jmp_cat",
  drinking_water_quality_jmp_cat_levels = c("surface_water", "unimproved", "limited",
    "basic", "safely_managed", "undefined"),
  sanitation_facility_jmp_cat = "wash_sanitation_facility_jmp_cat",
  sanitation_facility_jmp_cat_levels = c("open_defecation", "unimproved", "limited",
    "basic", "safely_managed", "undefined"),
  sanitation_facility_cat = "wash_sanitation_facility_cat",
  sanitation_facility_cat_levels = c("none", "unimproved", "improved", "undefined"),
  sanitation_facility_n_ind = "wash_sharing_sanitation_facility_n_ind",
  sanitation_facility_n_ind_levels = c("50_and_above", "20_to_49", "19_and_below"),
  handwashing_facility_jmp_cat = "wash_handwashing_facility_jmp_cat",
  handwashing_facility_jmp_cat_levels = c("no_facility", "limited", "basic", "undefined")
)

Arguments

df

A data frame.

setting

Column name for the setting.

setting_levels

Levels for the setting in that order: camp, rural, urban.

drinking_water_quantity

Column name for drinking water quantity.

drinking_water_quantity_levels

Levels for drinking water quantity.

drinking_water_quality_jmp_cat

Column name for drinking water quality JMP category.

drinking_water_quality_jmp_cat_levels

Levels for drinking water quality JMP category.

sanitation_facility_jmp_cat

Column name for sanitation facility JMP category.

sanitation_facility_jmp_cat_levels

Levels for sanitation facility JMP category.

sanitation_facility_cat

Column name for sanitation facility category.

sanitation_facility_cat_levels

Levels for sanitation facility category.

sanitation_facility_n_ind

Column name for number of individuals sharing a sanitation facility.

sanitation_facility_n_ind_levels

Levels for number of individuals sharing a sanitation facility.

handwashing_facility_jmp_cat

Column name for handwashing facility JMP category.

handwashing_facility_jmp_cat_levels

Levels for handwashing facility JMP category.