Drinking water source recoding

add_drinking_water_source_cat() recodes the types of water sources, add_drinking_water_time_cat() adds the categorical time to fetch water, and add_drinking_water_time_threshold_cat() according to a chosen threshold. Finally, add_drinking_water_quality_jmp_cat recodes the water source and time to fetch water into a joint JMP category.

Usage

add_drinking_water_source_cat(
  df,
  drinking_water_source = "wash_drinking_water_source",
  improved = c("piped_dwelling", "piped_compound", "piped_neighbour", "tap", "borehole",
    "protected_well", "protected_spring", "rainwater_collection", "tank_truck",
    "cart_tank", "kiosk", "bottled_water", "sachet_water"),
  unimproved = c("unprotected_well", "unprotected_spring"),
  surface_water = "surface_water",
  undefined = c("dnk", "pnta", "other")
)

add_drinking_water_time_cat(
  df,
  drinking_water_time_yn = "wash_drinking_water_time_yn",
  water_on_premises = "water_on_premises",
  number_minutes = "number_minutes",
  dnk = "dnk",
  undefined = "pnta",
  drinking_water_time_int = "wash_drinking_water_time_int",
  max = 600,
  drinking_water_time_sl = "wash_drinking_water_time_sl",
  sl_under_30_min = "under_30_min",
  sl_30min_1hr = "30min_1hr",
  sl_more_than_1hr = "more_than_1hr",
  sl_undefined = c("dnk", "pnta"),
  drinking_water_source = "wash_drinking_water_source",
  skipped_drinking_water_source_premises = "piped_dwelling",
  skipped_drinking_water_source_undefined = c("dnk", "pnta")
)

add_drinking_water_time_threshold_cat(
  df,
  drinking_water_time_cat = "wash_drinking_water_time_cat",
  premises = "premises",
  under_30min = c("under_30_min", "30min_1hr"),
  above_30min = "more_than_1hr",
  undefined = "undefined"
)

add_drinking_water_quality_jmp_cat(
  df,
  drinking_water_source_cat = "wash_drinking_water_source_cat",
  drinking_water_source_cat_levels = c("improved", "unimproved", "surface_water",
    "undefined"),
  drinking_water_time_30min_cat = "wash_drinking_water_time_30min_cat",
  drinking_water_time_30min_cat_levels = c("premises", "under_30min", "above_30min",
    "undefined")
)

Arguments

df

A data frame.

drinking_water_source

Component column: Water source types.

improved

Character vector of responses codes, such as "Protected well" or "Public tap", e.g., c("protected_well", "public_tap").

unimproved

Character vector of responses codes, such as "Unprotected well" or "Unprotected spring", e.g., c("unprotected_well", "unprotected_spring").

surface_water

Character vector of responses codes, such as "Lake" or "River, e.g., c("lake", "river").

undefined

Character vector of responses codes for undefined information, e.g. "Prefer not to answer".

drinking_water_time_yn

Component column: Time to fetch water, scoping question.

water_on_premises

Character vector of responses codes for water on premises.

number_minutes

Character vector of responses codes for number of minutes.

dnk

Character vector of responses codes for "Don't know".

drinking_water_time_int

Component column: Time to fetch water, integer.

max

Integer, the maximum value for the time to fetch water.

drinking_water_time_sl

Component column: Time to fetch water, simple choice.

sl_under_30_min

Response code for under 30 minutes.

sl_30min_1hr

Response code for 30 minutes to 1 hour.

sl_more_than_1hr

Response code for more than 1 hour.

sl_undefined

Character vector of responses codes for undefined information, e.g. "Don't know" or "Prefer not to answer".

skipped_drinking_water_source_premises

Character vector of responses codes for skipped water source on premises, e.g. "Piped into dwelling".

skipped_drinking_water_source_undefined

Character vector of responses codes for skipped water source undefined, e.g. "Don't know" or "Prefer not to answer".

drinking_water_time_cat

Component column: Time to fetch water, recoded categories.

premises

Character vector of responses codes for water on premises.

under_30min

Character vector of responses codes for under 30 min.

above_30min

Character vector of responses codes for above 30 min.

drinking_water_source_cat

Component column: Water source categories.

drinking_water_source_cat_levels

Character vector of water source categories in that order: Improved, Unimproved, Surface water, Undefined.

drinking_water_time_30min_cat

Component column: Time to fetch water, recoded categories.

drinking_water_time_30min_cat_levels

Character vector of time to fetch water categories in that order: Premises, Under 30 min, Above 30 min, Undefined.