Add functional domestic space tasks categories

Add functional domestic space tasks categories

Usage

add_fds_cannot_cat(
  df,
  fds_cooking = "snfi_fds_cooking",
  fds_cooking_cannot = "no_cannot",
  fds_cooking_can_issues = "yes_issues",
  fds_cooking_can_no_issues = "yes_no_issues",
  fds_cooking_no_need = "no_no_need",
  fds_cooking_undefined = c("pnta", "dnk"),
  fds_sleeping = "snfi_fds_sleeping",
  fds_sleeping_cannot = "no_cannot",
  fds_sleeping_can_issues = "yes_issues",
  fds_sleeping_can_no_issues = "yes_no_issues",
  fds_sleeping_undefined = c("pnta", "dnk"),
  fds_storing = "snfi_fds_storing",
  fds_storing_cannot = "no_cannot",
  fds_storing_can_issues = "yes_issues",
  fds_storing_can_no_issues = "yes_no_issues",
  fds_storing_undefined = c("pnta", "dnk"),
  fds_personal_hygiene = "snfi_fds_personal_hygiene",
  fds_personal_hygiene_cannot = "no_cannot",
  fds_personal_hygiene_can_issues = "yes_issues",
  fds_personal_hygiene_can_no_issues = "yes_no_issues",
  fds_personal_hygiene_undefined = c("pnta", "dnk"),
  lighting_source = "energy_lighting_source",
  lighting_source_none = "none",
  lighting_source_undefined = c("pnta", "dnk")
)

Arguments

df

A data frame

fds_cooking

Column name for cooking tasks

fds_cooking_cannot

Value for cannot perform cooking tasks

fds_cooking_can_issues

Value for can perform cooking tasks but with issues

fds_cooking_can_no_issues

Value for can perform cooking tasks without issues

fds_cooking_no_need

Value for no need to perform cooking tasks

fds_cooking_undefined

Vector of undefined responses

fds_sleeping

Column name for sleeping tasks

fds_sleeping_cannot

Value for cannot perform sleeping tasks

fds_sleeping_can_issues

Value for can perform sleeping tasks but with issues

fds_sleeping_can_no_issues

Value for can perform sleeping tasks without issues

fds_sleeping_undefined

Vector of undefined responses

fds_storing

Column name for storing tasks

fds_storing_cannot

Value for cannot perform storing tasks

fds_storing_can_issues

Value for can perform storing tasks but with issues

fds_storing_can_no_issues

Value for can perform storing tasks without issues

fds_storing_undefined

Vector of undefined responses

fds_personal_hygiene

Column name for personal hygiene tasks

fds_personal_hygiene_cannot

Value for cannot perform personal hygiene tasks

fds_personal_hygiene_can_issues

Value for can perform personal hygiene tasks but with issues

fds_personal_hygiene_can_no_issues

Value for can perform personal hygiene tasks without issues

fds_personal_hygiene_undefined

Vector of undefined responses

lighting_source

Column name for lighting source

lighting_source_none

Value for no lighting source

lighting_source_undefined

Vector of undefined responses